+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 8 September 2017
+- The ACCEPT-LANGUAGE header can now be used to set the locale on each request.
+  - Eg. 'ACCEPT-LANGUAGE: "fr-FR"'
+
 ## 5 September 2017
 - Added active_mpd_start_at, active_mpd_finish_at, active_mpd_monthly_goal fields to account_list object.
 - Added primary_appeal relationship to account_list object.